UPSTREAM: <714>: failing to obtain nodes should not be fatal
This is partial revert of upstream PR #714 which changed error handlig
so that failure to obtain nodes from apiserver would be fatal. We
should not increase RBAC permissions to grant node access.

This commit can be probably dropped in the next rebase because the
error might be changed again so it's not fatal.